+++ src/Components/Account/EditStudent.js
@@ -12,20 +12,19 @@ const Container = styled.div`
   row-gap: 0.625rem;
 `
 
-const EditStudent = ({ studentsNum }) => {
-  const [register, setRegister] = useState(false)
-  const onClickRegisterBtn = () => setRegister(prev => !prev)
+const EditStudent = ({ studentsNum, setRegisterPage, registerPage }) => {
+  const onClickRegisterBtn = () => {
+    if (registerPage) {
+      return
+    }
+    setRegisterPage("student")
+  }
   return (<Container>
     {studentsNum ? `${studentsNum}의 학생이 등록되어 있습니다.` : "등록된 학생이 없습니다."}
     <BtnContainer>
       <RegisterBtn onClick={onClickRegisterBtn}>{studentsNum ? "상세정보" : "등록하기"}</RegisterBtn>
       <DelBtn>학생정보 삭제하기</DelBtn>
     </BtnContainer>
-    {register && (studentsNum === 0 ?
-      <RegisterStudent />
-      :
-      <DetailStudent />
-    )}
   </Container>);
 }
